+++ test/Echo.IntegrationTests/Subject/Endpoint_ChargesExtra.cs
@@ -1,11 +1,11 @@
-﻿namespace Echo.IntegrationTests.Source
+﻿namespace Echo.IntegrationTests.Subject
 {
-    internal class Endpoint : IEndpoint
+    internal class Endpoint_ChargesExtra : IEndpoint
     {
         private readonly IBilling _billing;
         private readonly IProvider _provider;
 
-        public Endpoint(IBilling billing, IProvider provider)
+        public Endpoint_ChargesExtra(IBilling billing, IProvider provider)
         {
             _billing = billing;
             _provider = provider;
@@ -19,7 +19,7 @@
             });
             var paymentResponse = _billing.Charge(new PaymentRequest()
             {
-                Amount = quoteResponse.Price,
+                Amount = quoteResponse.Price + 1,
                 Instrument = request.Payment,
                 Payee = request.Customer,
             });
