+++ test/Echo.IntegrationTests/Subject/PaymentRequest.cs
@@ -1,4 +1,4 @@
-﻿namespace Echo.IntegrationTests.Source
+﻿namespace Echo.IntegrationTests.Subject
 {
     public class PaymentRequest
     {
