+++ test/Echo.IntegrationTests/Subject/ProvisioningRequest.cs
@@ -1,4 +1,4 @@
-﻿namespace Echo.IntegrationTests.Source
+﻿namespace Echo.IntegrationTests.Subject
 {
     public class ProvisioningRequest
     {
