+++ test/Echo.IntegrationTests/Subject/ServiceType.cs
@@ -1,4 +1,4 @@
-﻿namespace Echo.IntegrationTests.Source
+﻿namespace Echo.IntegrationTests.Subject
 {
     public enum ServiceType
     {
