+++ test/Echo.IntegrationTests/Subject/User.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Echo.IntegrationTests.Source
+namespace Echo.IntegrationTests.Subject
 {
     public class User
     {
